workshop-builder/requirements.txt
# Workshop Builder - OpenAI Codex Framework Dependencies
# Python package requirements for professional multi-agent orchestration

# =============================================================================
# CORE FRAMEWORK DEPENDENCIES
# =============================================================================

# Environment and Configuration Management
python-dotenv>=1.0.0          # Environment variable management
pydantic>=2.0.0               # Data validation and settings management

# =============================================================================
# AI API INTEGRATIONS
# =============================================================================

# Google Gemini Flash 2.5 Integration
google-generativeai>=0.3.0    # Official Google Generative AI client

# OpenAI API Integration (direct REST API calls)
openai>=1.52.0                # Official OpenAI Python client with latest features

# =============================================================================
# GITHUB INTEGRATION
# =============================================================================

# Professional GitHub API Integration
PyGithub>=1.59.0              # Comprehensive GitHub API client
gitpython>=3.1.0              # Git repository operations

# =============================================================================
# HTTP AND API CLIENTS
# =============================================================================

# HTTP Client Libraries
requests>=2.31.0              # HTTP library for API communications
httpx>=0.24.0                 # Modern async HTTP client
urllib3>=2.0.0                # HTTP client with connection pooling

# =============================================================================
# TEMPLATE AND CONTENT GENERATION
# =============================================================================

# Template Rendering
Jinja2>=3.1.0                 # Professional template engine
MarkupSafe>=2.1.0             # Safe string handling for templates

# Markdown Processing
markdown>=3.4.0               # Markdown processing and validation
python-markdown-math>=0.8     # Mathematical expressions in markdown

# =============================================================================
# CLI AND USER INTERFACE
# =============================================================================

# Command Line Interface
click>=8.1.0                  # Advanced CLI framework
rich>=13.0.0                  # Rich text and beautiful formatting
colorama>=0.4.6               # Cross-platform colored terminal text

# =============================================================================
# DATA PROCESSING AND VALIDATION
# =============================================================================

# JSON and Data Processing
jsonschema>=4.17.0            # JSON schema validation
pyyaml>=6.0                   # YAML processing for configuration
toml>=0.10.2                  # TOML configuration file support

# Date and Time Processing
python-dateutil>=2.8.0       # Extended date/time processing

# =============================================================================
# LOGGING AND MONITORING
# =============================================================================

# Enhanced Logging
structlog>=23.0.0             # Structured logging for professional output
loguru>=0.7.0                 # Advanced logging with better formatting

# =============================================================================
# SECURITY AND VALIDATION
# =============================================================================

# Security and Validation
cryptography>=41.0.0          # Cryptographic operations
certifi>=2023.0.0             # Certificate validation

# =============================================================================
# DEVELOPMENT AND TESTING DEPENDENCIES
# =============================================================================

# Testing Framework (for development)
pytest>=7.4.0                # Testing framework
pytest-asyncio>=0.21.0       # Async testing support
pytest-mock>=3.11.0          # Mocking for tests

# Code Quality (for development)
black>=23.0.0                 # Code formatting
isort>=5.12.0                 # Import sorting
flake8>=6.0.0                 # Code linting
mypy>=1.5.0                   # Type checking

# =============================================================================
# ASYNC AND CONCURRENCY SUPPORT
# =============================================================================

# Async Support
asyncio>=3.4.3                # Async I/O support (built-in, but explicit)
aiohttp>=3.8.0                # Async HTTP client
aiofiles>=23.0.0              # Async file operations

# =============================================================================
# SYSTEM AND PROCESS MANAGEMENT
# =============================================================================

# Process Management (for system monitoring)
psutil>=5.9.0                 # System and process utilities

# =============================================================================
# OPTIONAL PERFORMANCE ENHANCEMENTS
# =============================================================================

# Performance Optimizations
ujson>=5.8.0                  # Fast JSON encoder/decoder
orjson>=3.9.0                 # Fast JSON library

# =============================================================================
# COMPATIBILITY AND PLATFORM SUPPORT
# =============================================================================

# Cross-platform Support
pathlib2>=2.3.7; python_version<"3.4"  # Backport of pathlib
typing-extensions>=4.7.0     # Extended typing support

# =============================================================================
# NOTES
# =============================================================================

# Installation Instructions:
# 1. Create virtual environment: python -m venv .venv
# 2. Activate virtual environment: source .venv/bin/activate (Linux/Mac) or .venv\Scripts\activate (Windows)
# 3. Install dependencies: pip install -r requirements.txt
# 4. Verify installation: python -c "import google.generativeai, openai, github; print('✓ All dependencies installed')"

# Version Compatibility:
# - Python 3.10+ required for OpenAI API compatibility
# - All packages tested with Python 3.10, 3.11, and 3.12
# - Some packages may have newer versions available; update as needed

# Security Considerations:
# - All packages are from trusted sources (PyPI official packages)
# - Regular security updates recommended
# - Use pip-audit for security vulnerability scanning: pip install pip-audit && pip-audit

# Performance Notes:
# - ujson and orjson provide significant JSON processing performance improvements
# - aiohttp enables async HTTP operations for better concurrency
# - psutil provides system monitoring capabilities for resource management